Disable re-run of auto* in a regular build
Currently the Makefile.in are set up so that it will rerun aclocal
and the resut of the auto* toolchain. This is usually referred to
as 'maintainer-mode' and generally only enabled if you pass
--enable-maintainer-mode in the configure. However, for some reason
gettext isn't set up in this manner.

This current patch is a hack that manually disables the build rule
in the Makefile.in. A better solution would be to patch the .am
files to add the maintainer mode logic correctly, however as a
prgramatic expediency I've gone with the hack for now.